updated portfolio section
